feat: обновление типов данных для totalAmount и оптимизация обработки цен в корзине
Изменены типы данных для полей totalAmount в моделях Cart и Order с Int на Decimal для повышения точности расчетов. Оптимизированы функции обработки цен в компонентах и API маршрутах, включая использование числовых типов для корректных вычислений. Эти изменения улучшают точность данных и производительность приложения.

+++ src/app/api/cart/route.ts
@@ -138,12 +138,12 @@ export async function POST(req: NextRequest) {
 					"totalAmount" = COALESCE((
 						SELECT SUM(
 							(pi.price + COALESCE(
-								(SELECT SUM(ing.price)::int 
-								 FROM "_CartItemToIngredient" m
-								 JOIN "Ingredient" ing ON ing.id = m."B"
-								 WHERE m."A" = ci.id), 
+								(SELECT SUM(ing.price)::numeric   
+								FROM "_CartItemToIngredient" m
+								JOIN "Ingredient" ing ON ing.id = m."B"
+								WHERE m."A" = ci.id), 
 							0)) * ci.quantity
-						)::int
+						)::numeric 
 						FROM "CartItem" ci
 						JOIN "ProductItem" pi ON pi.id = ci."productItemId"
 						WHERE ci."cartId" = c.id
